Move more tmpfiles code into tmpfiles module
Before we kind of viewed the "translate tmpfiles in package import"
as the core flow, but I think that needs to be a peer of
"postprocess tmpfiles".

Move all the tmpfiles code from importer.rs into tmpfiles.rs.

No functional changes.
