+++ rust/src/tmpfiles.rs
@@ -6,6 +6,7 @@
 use crate::cxxrsutil::*;
 use crate::ffiutil;
 
+use anyhow::bail;
 use anyhow::{Context, Result};
 use camino::Utf8Path;
 use camino::Utf8PathBuf;
@@ -16,8 +17,10 @@ use cap_std::io_lifetimes::AsFilelike;
 use cap_std_ext::dirext::CapStdExtDirExt;
 use fn_error_context::context;
 use ostree_ext::gio;
+use ostree_ext::gio::FileInfo;
 use ostree_ext::gio::FileType;
 use ostree_ext::prelude::CancellableExtManual;
+use std::borrow::Cow;
 use std::collections::BTreeMap;
 use std::collections::BTreeSet;
 use std::fmt::Write;
@@ -28,6 +31,67 @@ use std::path::PathBuf;
 const TMPFILESD: &str = "usr/lib/tmpfiles.d";
 const RPMOSTREE_TMPFILESD: &str = "usr/lib/rpm-ostree/tmpfiles.d";
 const AUTOVAR_PATH: &str = "rpm-ostree-autovar.conf";
+
+fn fix_tmpfiles_path(abs_path: std::borrow::Cow<str>) -> Cow<str> {
+    let mut tweaked_path = abs_path;
+
+    // systemd-tmpfiles complains loudly about writing to /var/run;
+    // ideally, all of the packages get fixed for this but...eh.
+    if tweaked_path.starts_with("/var/run/") {
+        let trimmed = tweaked_path.trim_start_matches("/var");
+        tweaked_path = Cow::Owned(trimmed.to_string());
+    }
+
+    // Handle file paths with spaces and other chars;
+    // see https://github.com/coreos/rpm-ostree/issues/2029 */
+    crate::utils::shellsafe_quote(tweaked_path)
+}
+
+/// Translate a filepath entry to an equivalent tmpfiles.d line.
+#[context("Translating {}", abs_path)]
+pub(crate) fn translate_to_tmpfiles_d(
+    abs_path: &str,
+    file_info: &FileInfo,
+    username: &str,
+    groupname: &str,
+) -> Result<String> {
+    let mut bufwr = String::new();
+
+    let path_type = file_info.file_type();
+    let filetype_char = match path_type {
+        FileType::Directory => 'd',
+        FileType::Regular => 'f',
+        FileType::SymbolicLink => 'L',
+        x => bail!("path '{}' has invalid type: {:?}", abs_path, x),
+    };
+    let fixed_path = fix_tmpfiles_path(Cow::Borrowed(abs_path));
+    write!(&mut bufwr, "{} {}", filetype_char, fixed_path)?;
+
+    if path_type == FileType::SymbolicLink {
+        let link_target = file_info
+            .symlink_target()
+            .ok_or_else(|| anyhow::anyhow!("missing symlink target"))?;
+        let link_target = link_target
+            .to_str()
+            .ok_or_else(|| anyhow::anyhow!("Invalid non-UTF8 symlink: {link_target:?}"))?;
+        write!(&mut bufwr, " - - - - {}", link_target)?;
+    } else {
+        let mode = file_info.attribute_uint32("unix::mode") & !libc::S_IFMT;
+        write!(&mut bufwr, " {:04o} {} {} - -", mode, username, groupname)?;
+    };
+
+    Ok(bufwr)
+}
+
+pub fn tmpfiles_translate(
+    abs_path: &str,
+    file_info: &crate::FFIGFileInfo,
+    username: &str,
+    groupname: &str,
+) -> CxxResult<String> {
+    let entry = translate_to_tmpfiles_d(abs_path, &file_info.glib_reborrow(), username, groupname)?;
+    Ok(entry)
+}
 
 #[context("Converting /var to tmpfiles.d")]
 pub(crate) fn var_to_tmpfiles(rootfs: &Dir, cancellable: Option<&gio::Cancellable>) -> Result<()> {
@@ -164,12 +228,7 @@ fn convert_path_to_tmpfiles_d_recurse(
             let abs_path = Utf8Path::new("/").join(&full_path);
             let username = pwdb.lookup_user(meta.uid())?;
             let groupname = pwdb.lookup_group(meta.gid())?;
-            crate::importer::translate_to_tmpfiles_d(
-                abs_path.as_str(),
-                &file_info,
-                &username,
-                &groupname,
-            )?
+            translate_to_tmpfiles_d(abs_path.as_str(), &file_info, &username, &groupname)?
         };
         out_entries.insert(entry);
 
@@ -453,5 +512,64 @@ q /var/tmp 1777 root root 30d
             "f /var/lib/nfs/etab 0770 test-user test-group - -",
         ];
         assert_eq!(entries, expected, "{:#?}", entries);
+    }
+
+    #[test]
+    fn test_fix_tmpfiles_path() {
+        let intact_cases = vec!["/", "/var", "/var/foo", "/run/foo"];
+        for entry in intact_cases {
+            let output = fix_tmpfiles_path(Cow::Borrowed(entry));
+            assert_eq!(output, entry);
+        }
+
+        let quoting_cases = maplit::btreemap! {
+            "/var/foo bar" => r#"'/var/foo bar'"#,
+            "/var/run/" => "/run/",
+            "/var/run/foo bar" => r#"'/run/foo bar'"#,
+        };
+        for (input, expected) in quoting_cases {
+            let output = fix_tmpfiles_path(Cow::Borrowed(input));
+            assert_eq!(output, expected);
+        }
+    }
+
+    #[test]
+    fn test_translate_to_tmpfiles_d() {
+        let path = r#"/var/foo bar"#;
+        let username = "testuser";
+        let groupname = "testgroup";
+        {
+            // Directory
+            let file_info = FileInfo::new();
+            file_info.set_file_type(FileType::Directory);
+            file_info.set_attribute_uint32("unix::mode", 0o721);
+            let out = translate_to_tmpfiles_d(path, &file_info, username, groupname).unwrap();
+            let expected = r#"d '/var/foo bar' 0721 testuser testgroup - -"#;
+            assert_eq!(out, expected);
+        }
+        {
+            // Symlink
+            let file_info = FileInfo::new();
+            file_info.set_file_type(FileType::SymbolicLink);
+            file_info.set_symlink_target("/mytarget");
+            let out = translate_to_tmpfiles_d(path, &file_info, username, groupname).unwrap();
+            let expected = r#"L '/var/foo bar' - - - - /mytarget"#;
+            assert_eq!(out, expected);
+        }
+        {
+            // File
+            let file_info = FileInfo::new();
+            file_info.set_file_type(FileType::Regular);
+            file_info.set_attribute_uint32("unix::mode", 0o123);
+            let out = translate_to_tmpfiles_d(path, &file_info, username, groupname).unwrap();
+            let expected = r#"f '/var/foo bar' 0123 testuser testgroup - -"#;
+            assert_eq!(out, expected);
+        }
+        {
+            // Other unsupported
+            let file_info = FileInfo::new();
+            file_info.set_file_type(FileType::Unknown);
+            translate_to_tmpfiles_d(path, &file_info, username, groupname).unwrap_err();
+        }
     }
 }
